% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{captcha_download}
\alias{captcha_download}
\alias{captcha_download_tjrs}
\alias{captcha_download_tjmg}
\alias{captcha_download_trt}
\alias{captcha_download_rfb}
\alias{captcha_download_tjrj}
\title{Download captcha from a site}
\usage{
captcha_download(url, n = 1, dest = ".", secure = FALSE,
  .default = "jpeg")

captcha_download_tjrs(n = 1, dest = ".")

captcha_download_tjmg(n = 1, dest = ".")

captcha_download_trt(n = 1, dest = ".")

captcha_download_rfb(n = 1, dest = ".")

captcha_download_tjrj(n = 1, dest = ".")
}
\arguments{
\item{url}{url to download. Must be a valid url to download the captcha.
One may check whether the url is working or not running
\code{\link[httr]{BROWSE}(url)}.}

\item{n}{total number of captchas to download. Defaults to one.}

\item{dest}{captcha file destination. Defaults to current directory.}

\item{secure}{use option \code{ssl_verifypeer = TRUE}? Defaults to \code{FALSE}.}

\item{.default}{image extension if not extracted automatically from
  \code{httr::GET(url, ...)$headers[["content-type"]]}.
  Normally "jpeg" or "png". Defaults to "jpeg".

All downloads have a timeout of three seconds to run. If the site you are
accessing blocks IP after multiple calls, consider creating a loop an using
\code{\link[base]{Sys.sleep}} to wait new calls.}
}
\description{
\code{captcha_download} is a generic function to download captchas from any
site, given the url captcha. It is possible to download many captchas at
onde with the \code{n} option. \code{captcha_download_*} are aliases for
some known captchas of public services in Brazil.
}
\examples{
\dontrun{
  captcha_download_rfb()
  captcha_download_trt(10, 'my/folder')
  captcha_download("https://goo.gl/sqTQ4Z")
}
}
