% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{captcha_prep_options}
\alias{captcha_prep_options}
\title{Options to prepare captcha}
\usage{
captcha_prep_options(img_size = c(24L, 72L))
}
\arguments{
\item{img_size}{image size}
}
\description{
Options to prepare captcha
}
