% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{captcha_prepare_dataset}
\alias{captcha_prepare_dataset}
\title{Prepare dataset to run keras models}
\usage{
captcha_prepare_dataset(dir, n_train, vocab, batch_size, prep_options)
}
\arguments{
\item{dir}{directory}

\item{n_train}{number of training examples}

\item{vocab}{captcha vocabulary. If missing will be inferred from dataset}

\item{batch_size}{size of the batch}

\item{prep_options}{options to prepare captcha, returned
by \code{\link[=captcha_prep_options]{captcha_prep_options()}}}
}
\description{
Prepare dataset to run keras models
}
