% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify.captcha}
\alias{classify.captcha}
\title{Classify captcha}
\usage{
\method{classify}{captcha}(x, dest, answer = NULL, ...)
}
\arguments{
\item{x}{object}

\item{dest}{destination}

\item{answer}{answer}

\item{...}{other}
}
\description{
Classify captcha
}
