% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{decrypt}
\alias{decrypt}
\title{Break captchas}
\usage{
decrypt(captcha, model)
}
\arguments{
\item{captcha}{The paths to one or more captchas or a list of
captchas read with \code{\link[=read_captcha]{read_captcha()}}}

\item{model}{Either a model, the path to a model or the name of a
known model (see \code{\link[=train_model]{train_model()}} and \code{\link[=load_model]{load_model()}})}
}
\description{
Given one or more captchas and a model, returns the
text that is supposed to break the captchas.
}
