% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{decrypt_}
\alias{decrypt_}
\title{Break a captcha}
\usage{
decrypt_(captcha, model)
}
\arguments{
\item{captcha}{A captcha read with \code{\link[=read_captcha]{read_captcha()}}}

\item{model}{A model loaded with \code{\link[=load_model]{load_model()}}}
}
\description{
Break a captcha
}
