% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download captcha from any site}
\usage{
download(url, dest = NULL, secure = FALSE)
}
\arguments{
\item{url}{url to download}

\item{dest}{destination}

\item{secure}{use ssl verifypeer as FALSE}
}
\description{
Generic download function
}
