% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_captcha}
\alias{download_captcha}
\title{Download captchas from a URL}
\usage{
download_captcha(url, n = 1, path = ".", secure = FALSE, ext = ".jpeg")
}
\arguments{
\item{url}{URL from which to download captcha or the name of a known source
(\code{"tjrs"}, \code{"tjmg"}, \code{"tjrj"}, \code{"trt"} or \code{"rfb"})}

\item{n}{Total number of captchas to download}

\item{path}{Folder where to save downloaded captchas}

\item{secure}{Whether or not to use \code{ssl_verifypeer = TRUE} (see
\code{\link[httr:GET]{httr::GET()}})}

\item{ext}{Default image extension if not able to extract it automatically}
}
\value{
A vector with the paths to the downloaded files
}
\description{
This is a generic function to download captchas from any
webssite given a captcha's URL. \code{download_captcha_*()} are aliases for
some known captchas of public services in Brazil.
}
\section{IP blocking}{

All downloads have a timeout of three seconds to run.
If the website you are accessing blocks IP after multiple calls, consider
creating a loop an using \code{\link[base:Sys.sleep]{base::Sys.sleep()}} to wait for new calls.
}

