% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_captcha_}
\alias{download_captcha_}
\title{Download captcha from a URL}
\usage{
download_captcha_(url, path, secure, ext, timeout)
}
\arguments{
\item{url}{URL from which to download captcha or the name of a known source
(\code{"tjrs"}, \code{"tjmg"}, \code{"tjrj"}, \code{"trt"} or \code{"rfb"})}

\item{path}{Folder where to save downloaded captcha}

\item{secure}{Whether or not to use \code{ssl_verifypeer = TRUE} (see
\code{\link[httr:GET]{httr::GET()}})}

\item{ext}{Default image extension if not able to extract it automatically}

\item{timeout}{Timeout for captcha download (passed on to \code{\link[httr:timeout]{httr::timeout()}})}
}
\description{
Download captcha from a URL
}
