% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_answer}
\alias{get_answer}
\title{Get answer from filename and build answer matrix}
\usage{
get_answer(file, vocab)
}
\arguments{
\item{file}{Path to a captcha with its answer separated by and
underscore ('_') in the filename}

\item{vocab}{Character vector with all possible values in the captcha. If
not specified, answers won't be transformed to the matrix format.}
}
\description{
Get answer from filename and build answer matrix
}
