% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{join_answers}
\alias{join_answers}
\title{Join answer matrixes}
\usage{
join_answers(ys)
}
\arguments{
\item{ys}{A list of answer matrixes (captcha$y)}
}
\description{
Given a list of answer matrixes, joins them in the
shape expected by the model of \code{\link[=train_model]{train_model()}}.
}
