% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{join_captchas}
\alias{join_captchas}
\title{Join captcha objects}
\usage{
join_captchas(captchas)
}
\arguments{
\item{captchas}{A list of answered captchas (read by
\code{\link[=read_captcha]{read_captcha()}})}
}
\description{
Given a list of captcha objetcs, joins them in the
shape expected by the model of \code{\link[=train_model]{train_model()}}.
}
