% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{load_image}
\alias{load_image}
\title{Read PNG or JPG}
\usage{
load_image(path, ...)
}
\arguments{
\item{path}{Path to image}

\item{...}{Other arguments passed on to \code{\link[png:readPNG]{png::readPNG()}}
or \code{\link[jpeg:readJPEG]{jpeg::readJPEG()}}}
}
\description{
Read PNG or JPG
}
