% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.captcha}
\alias{model.captcha}
\title{Model image captcha using keras}
\usage{
\method{model}{captcha}(prepared_data, n_test = prepared_data$n * 0.1,
  n_epochs = 30, n_units = 256, batch_size = 64, save_model = TRUE,
  save_path = "model.hdf5", verbose = TRUE)
}
\arguments{
\item{prepared_data}{prepared data}

\item{n_test}{test data size}

\item{n_epochs}{maximum number of epochs}

\item{n_units}{number of units to consider in dense layer}

\item{batch_size}{number of observations to consider in each iteration}

\item{save_model}{logical. save model on disk?}

\item{save_path}{path to save model}

\item{verbose}{print model specification}
}
\description{
Tries to adjust an automatic model using keras
}
