% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.audio_captcha}
\alias{plot.audio_captcha}
\title{Plot captcha}
\usage{
\method{plot}{audio_captcha}(x, y, ...)
}
\arguments{
\item{x}{object}

\item{y}{-}

\item{...}{other}
}
\description{
Plot captcha
}
