% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.captcha}
\alias{plot.captcha}
\title{Plot a captcha}
\usage{
\method{plot}{captcha}(x, y, ...)
}
\arguments{
\item{x}{Captcha object read with \code{\link[=read_captcha]{read_captcha()}}}

\item{y}{\itemize{
\item 
}}

\item{...}{Other arguments passed on to \code{\link[graphics:plot.default]{graphics::plot()}}}
}
\description{
Plot a captcha
}
