% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tjrs}
\alias{predict.tjrs}
\title{Predizer os numeros do arquivo}
\usage{
\method{predict}{tjrs}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other}
}
\description{
Predizer os numeros do arquivo
}
