% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.captcha}
\alias{prepare.captcha}
\title{Prepare captchas}
\usage{
\method{prepare}{captcha}(arqs, only_x = FALSE)
}
\arguments{
\item{arqs}{arqs read}

\item{only_x}{boolean. Is the answers present on file names?}
}
\description{
Prepare answar and features for modeling. Expect '_' as the answer separator.
}
