% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.raw}
\alias{prepare.raw}
\title{Predict captcha}
\usage{
\method{prepare}{raw}(arqs, only_x = TRUE)
}
\arguments{
\item{arqs}{object}

\item{only_x}{boolean. Is the answers present on file names?}
}
\description{
Predict captcha
}
