% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Predict captcha}
\usage{
prepare(arqs, only_x = FALSE)
}
\arguments{
\item{arqs}{object}

\item{only_x}{boolean. Is the answers present on file names?}
}
\description{
Predict captcha
}
