% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.captcha}
\alias{print.captcha}
\title{Plot a captcha}
\usage{
\method{print}{captcha}(x, ...)
}
\arguments{
\item{x}{Captcha object read with \code{\link[=read_captcha]{read_captcha()}}}

\item{...}{Other arguments passed on to \code{\link[graphics:plot]{graphics::plot()}}}
}
\description{
Plot a captcha
}
