% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{read_model}
\alias{read_model}
\title{Read a model from \code{decryptrModels}}
\usage{
read_model(model)
}
\arguments{
\item{model}{The name of the model (\code{"rbf"}, \code{"trt"}, \code{"tjmg"} or
\code{"esaj"})}
}
\description{
Acesses the models saved in \code{decryptrModels:::models} and
makes them useful for cracking captchas locally
}
