% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train_model}
\alias{train_model}
\title{Craete a model to break image captchas using keras}
\usage{
train_model(data, frac_test = 0.1, n_epochs = 30, n_units = 256,
  batch_size = 64, path = "./model.hdf5", verbose = TRUE)
}
\arguments{
\item{data}{A list of previously-classified captchas read with
\code{read_captcha(..., ans_in_path = TRUE)} (see \code{\link[=classify]{classify()}})}

\item{frac_test}{Fraction of data used for testing}

\item{n_epochs}{Maximum number of epochs}

\item{n_units}{Number of units to consider in dense layer}

\item{batch_size}{Number of observations to consider in each iteration}

\item{path}{Path where to save model (if \code{NULL} won't save it)}

\item{verbose}{Print model specification}
}
\description{
This function tries to adjust an automatic model for
breaking captchas using the \code{rstudio/keras} package. It creates a
\code{\link[keras:keras_model_sequential]{keras::keras_model_sequential()}} and learns from the inputed
labeled data.
}
