% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify.captcha}
\alias{classify.captcha}
\title{Classify captchas}
\usage{
\method{classify}{captcha}(captcha, dest = dirname(captcha)[1],
  answer = NULL, ...)
}
\arguments{
\item{captcha}{object}

\item{dest}{destination}

\item{answer}{answer}

\item{...}{other}
}
\description{
Classify captchas
}
