% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{decrypt}
\alias{decrypt}
\title{Break a captcha}
\usage{
decrypt(captcha, model)
}
\arguments{
\item{captcha}{The path to a captcha or a captcha read with
\code{\link[=read_captcha]{read_captcha()}}}

\item{model}{The name of a model or a model from \code{decryptrModels}
(see \code{\link[=read_model]{read_model()}})}
}
\description{
Given a captcha and a model, returns the text that
is supposed to break the captcha.
}
