% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{download_tjrs}
\alias{download_tjmg}
\alias{download_trt}
\alias{download_rfb}
\alias{download_saj}
\alias{download_tjrj}
\title{Download captcha from any site}
\usage{
download(url, dest, n, secure = FALSE, type = "jpeg")

download_tjrs(dest = NULL, n = 1)

download_tjmg(dest = NULL, n = 1)

download_trt(dest = NULL, n = 1)

download_rfb(dest = NULL, n = 1)

download_saj(dest = NULL, n = 1)

download_tjrj(dest = NULL, n = 1)
}
\arguments{
\item{url}{url to download}

\item{dest}{destination}

\item{n}{number of captchas to download}

\item{secure}{use ssl verifypeer as FALSE}

\item{type}{image extension. Normally 'jpeg' or 'png'}
}
\description{
Generic download function
}
