% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download captcha from any site}
\usage{
download(url, dest = NULL, secure = FALSE, type = "jpeg")
}
\arguments{
\item{url}{url to download}

\item{dest}{destination}

\item{secure}{use ssl verifypeer as FALSE}

\item{type}{image extension. Normally 'jpeg' or 'png'}
}
\description{
Generic download function
}
