% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_answer}
\alias{get_answer}
\title{Get answer from filename and build answer matrix}
\usage{
get_answer(path)
}
\arguments{
\item{path}{Path to a captcha with its answer separated by and
underscore ('_') in the filename}
}
\description{
Get answer from filename and build answer matrix
}
