% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{load_image}
\alias{load_image}
\title{Read PNG or JPG}
\usage{
load_image(file, ...)
}
\arguments{
\item{file}{Path to image}

\item{...}{Other arguments passed on to \code{\link[magick:image_read]{magick::image_read()}}}
}
\description{
Read PNG or JPG
}
