% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_model}
\alias{load_model}
\title{Load a captcha-breaking model}
\usage{
load_model(model, labs = c(0:9, base::letters))
}
\arguments{
\item{model}{Either the path to a ".hdf5" file or the name of a known model
(\code{"rfb"}, \code{"trt"}, \code{"tjmg"} or \code{"esaj"})}

\item{labs}{A character vector with all the labels the model could possibly
output (necessary when loading a homemade model)}
}
\description{
This function can either access the models saved in
\code{decryptrModels:::models} or in a local file create by the user with
\code{\link[=train_model]{train_model()}}.
}
