% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.captcha}
\alias{model.captcha}
\title{Model image captcha using keras}
\usage{
\method{model}{captcha}(prepared, epochs = 100)
}
\arguments{
\item{prepared}{prepared data}

\item{epochs}{maximum number of epochs

#@import keras}
}
\description{
Model image captcha using keras
}
