% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.captcha}
\alias{predict.captcha}
\title{Predizer os numeros do arquivo}
\usage{
\method{predict}{captcha}(object, ...)
}
\arguments{
\item{object}{keras model}

\item{...}{captcha file}
}
\description{
Predizer os numeros do arquivo
}
