% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.captcha}
\alias{print.captcha}
\title{Print information about a captcha}
\usage{
\method{print}{captcha}(x, ...)
}
\arguments{
\item{x}{Captcha object read with \code{\link[=read_captcha]{read_captcha()}}}

\item{...}{\itemize{
\item 
}}
}
\description{
Print information about a captcha
}
