% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.model}
\alias{print.model}
\title{Print information about a model}
\usage{
\method{print}{model}(x, ...)
}
\arguments{
\item{x}{Model object read with \code{\link[=read_model]{read_model()}}}

\item{...}{\itemize{
\item 
}}
}
\description{
Print information about a model
}
