% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_captcha}
\alias{read_captcha}
\title{Read captcha files}
\usage{
read_captcha(file, ans_in_path = FALSE, vocab = NULL)
}
\arguments{
\item{file}{Paths to one or more captcha images}

\item{ans_in_path}{Whether or not the answers to the captchas are already
in the paths to the files (separated by and underscore in the filename)}

\item{vocab}{Character vector with all possible values in the captcha. If
not specified, answers won't be transformed to the matrix format.}
}
\value{
A list of captcha objects
}
\description{
Given the paths to one or more files, reads and converts
them into a \code{captcha} list that can be used for classification or
decryption. If \code{ans_in_path = TRUE}, will take the answer for the
captchas from their filenames and get them ready for modeling.
}
